refactor: Migration room modal to TS
